+++ js/main.js
@@ -1,6 +1,8 @@
 import { eventCache} from './modules/Proxy.js';
 import { renderEvents } from './utils/events.js';
 import { initializeButtonListeners } from './modules/buttonsChange.js';
+import { eventListeners } from './utils/buttonsTarget.js';
+
 
 document.querySelector('.tab-nav').addEventListener('click', async function(event) {
   if (event.target.classList.contains('category-button')) {
@@ -12,3 +14,5 @@ document.querySelector('.tab-nav').addEventListener('click', async function(even
 
 initializeButtonListeners();
 
+
+eventListeners();
