+++ myAccount.html
@@ -11,9 +11,9 @@
   <div>
     <ul class="tab-nav">
       <li><button type="button"><img src="svg/arrow-leftcom.svg" alt="arrow left" class="arrowLeft">Back to events</button></li>
-      <li><button class="category-button" data-category="music">Favorites</button></li>
-      <li><button class="category-button" data-category="sports">Interested</button></li>
-      <li><button class="category-button" data-category="business">Going</button></li>
+      <li><button class="category-button">Favorites</button></li>
+      <li><button class="category-button">Interested</button></li>
+      <li><button class="category-button">Going</button></li>
     </ul>
     <div class="event-grid"></div>
   </div>
